+++ database/migrations/2021_07_19_195709_create_article_incomes_table.php
@@ -16,8 +16,8 @@ class CreateArticleIncomesTable extends Migration
         Schema::create('article_incomes', function (Blueprint $table) {
             $table->id();
             $table->float('quantity');
-            $table->float('unit_price',11,2);
-            $table->float('total_price',11,2);
+            $table->decimal('unit_price',11,2);
+            $table->decimal('total_price',11,2);
             $table->foreignId('income_id')->constrained();
             $table->foreignId('article_id')->constrained();
             $table->timestamps();
